Design-informed size factor estimation

Design-informed size factor estimation

Usage

disize(
  design_formula,
  counts,
  metadata,
  batch_name,
  obs_name = "obs_id",
  n_feats = min(10000L, ncol(counts)),
  n_iters = 10000L,
  rel_tol = 1000,
  init_alpha = 1e-08,
  n_threads = 1L,
  n_tries = 5L,
  verbose = 3L
)

Arguments

design_formula

The formula describing the experimental design.

counts

A (observation x feature) count matrix.

metadata

A dataframe containing observation-level metadata.

batch_name

The identifier for the batch column in 'metadata'.

obs_name

The identifier for the observation column in 'metadata'.

n_feats

The number of features used during estimation.

n_iters

The number of iterations used for estimation.

rel_tol

The relative tolerance used for convergence.

init_alpha

The initial step-size for the optimizer, lower values can sometimes make it easier to estimate size factors for more complex designs.

n_threads

The number of threads to use for parallel processing.

n_tries

The maximum number of times to try fitting.

verbose

The verbosity level (1: only errors, 2: also allows warnings,3: also allows messages, 4: also prints additional output useful for debugging).

Value

A named numeric vector containing the size factor estimates.