Design-informed size factor estimation

Design-informed size factor estimation

Usage

disize(
  design_formula,
  counts,
  metadata,
  batch_name = "batch_id",
  obs_name = "obs_id",
  n_feats = min(10000, ncol(counts)),
  n_subset = 50,
  n_iters = "auto",
  n_threads = parallel::detectCores()/2,
  init_alpha = 1e-05,
  verbose = 3
)

Arguments

design_formula

The formula describing the experimental design.

counts

A (obsservation x feature) count matrix.

metadata

A dataframe containing sample metadata.

batch_name

The identifier for the batch column in 'metadata', defaults to "batch_id".

obs_name

The identifier for the observation column in 'metadata', defaults to "obs_id".

n_feats

The number of features used during estimation, defaults to automatically choosing based on a heuristic.

n_subset

The number of observations per experimental unit used during estimation, defaults to 50.

n_iters

The number of iterations used for a single optimization pass.

n_threads

The number of threads to use for parallel processing.

verbose

The verbosity level.

Value

A named numeric vector containing the size factor point estimates.